requirements.txt
snowflake-connector-python>=2.7.0
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.0
joblib>=1.1.0
python-dotenv>=1.0.0
pyarrow<19.0.0 